Add deferred loading and prefetching (#734)
This pull request implements deferred loading for repositories on the
index page. It updates the `loader` function to use the `defer` function
from `@remix-run/node` and adds `Suspense` components to handle the
loading states. This improves the performance of the index page by
delaying the loading of data until it is actually needed.

Furthermore, links on the index page are prefetched when you hover the
links, and the link back home from the repository view is prefeched on
render, making navigation back instant.