Allow multiple arguments to log methods
This resolves an issue with the logging module not conforming to the same API as native  console methods like console.log. This commit adds support for passing multiple arguments. The method will then automatically concatenate these arguments in the logged message.

This resolves a type error in InstanceManager, where it was expected that this API was supported:

```typescript
log.warn("getOrCreateInstance", repo, branch, path)
```